… hack/dind-cluster-v1.11.sh