@TravisBuddy 2018-07-24
## Travis tests have failed

Hey @k82cn,
Please read the following log in order to understand the failure reason.
It'll be awesome if you fix what's wrong and commit the changes.


### 1st Build


<details>
  <summary>
    <strong>
     make e2e
    </strong>
  </summary>

```
mkdir -p _output/bin
go build -o _output/bin/kar-scheduler ./cmd/kar-scheduler/
go build -o _output/bin/kar-controllers ./cmd/kar-controllers/
go build -o _output/bin/karcli ./cmd/karcli
hack/dind-cluster-v1.11.sh up
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
* Making sure DIND image is up to date 
Warning: failed to get default registry endpoint from daemon (Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?). Using system default: https://index.docker.io/v1/
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
make: *** [e2e] Error 1
```

</details>