@TravisBuddy 2018-07-24
## Travis tests have failed

Hey @k82cn,
Please read the following log in order to understand the failure reason.
It'll be awesome if you fix what's wrong and commit the changes.


### 1st Build


<details>
  <summary>
    <strong>
     make run-test
    </strong>
  </summary>

```
hack/make-rules/test.sh  
Running tests for APIVersion: v1,admissionregistration.k8s.io/v1alpha1,admission.k8s.io/v1alpha1,apps/v1beta1,apps/v1beta2,authentication.k8s.io/v1,authentication.k8s.io/v1beta1,authorization.k8s.io/v1,authorization.k8s.io/v1beta1,autoscaling/v1,autoscaling/v2beta1,batch/v1,batch/v1beta1,batch/v2alpha1,certificates.k8s.io/v1beta1,extensions/v1beta1,imagepolicy.k8s.io/v1alpha1,networking.k8s.io/v1,policy/v1beta1,rbac.authorization.k8s.io/v1,rbac.authorization.k8s.io/v1beta1,rbac.authorization.k8s.io/v1alpha1,scheduling.k8s.io/v1alpha1,settings.k8s.io/v1alpha1,storage.k8s.io/v1beta1,storage.k8s.io/v1,
+++ [0724 06:30:29] Running tests without code coverage
ok  	github.com/kubernetes-incubator/kube-arbitrator/pkg/scheduler/actions/allocate	0.048s
ok  	github.com/kubernetes-incubator/kube-arbitrator/pkg/scheduler/actions/decorate	0.011s
go build -o _output/bin/kar-controllers ./cmd/kar-controllers/
go build -o _output/bin/karcli ./cmd/karcli
hack/dind-cluster-v1.11.sh up
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
* Making sure DIND image is up to date 
Warning: failed to get default registry endpoint from daemon (Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?). Using system default: https://index.docker.io/v1/
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?
make: *** [e2e] Error 1
```

</details>


<details>
  <summary>
    <strong>
     make e2e
    </strong>
  </summary>

```
Done. Your build exited with 1.
```

</details>