@TravisBuddy 2018-07-24
## Travis tests have failed

Hey @k82cn,
Please read the following log in order to understand the failure reason.
It'll be awesome if you fix what's wrong and commit the changes.


### 1st Build


<details>
  <summary>
    <strong>
     make e2e
    </strong>
  </summary>

```
mkdir -p _output/bin
go build -o _output/bin/kar-scheduler ./cmd/kar-scheduler/
go build -o _output/bin/kar-controllers ./cmd/kar-controllers/
go build -o _output/bin/karcli ./cmd/karcli
hack/dind-cluster-v1.11.sh up
* Making sure DIND image is up to date 
v1.11: Pulling from mirantis/kubeadm-dind-cluster




















Digest: sha256:6041138151c4737db10441aeaf7f356f6271812109b54d0936a837846ed13ecb
Status: Downloaded newer image for mirantis/kubeadm-dind-cluster:v1.11
/home/travis/.kubeadm-dind-cluster/kubectl-v1.11.0: OK
* Starting DIND container: kube-master
* Running kubeadm: init --config /etc/kubeadm.conf --skip-preflight-checks
Initializing machine ID from random generator.
Synchronizing state of docker.service with SysV service script with /lib/systemd/systemd-sysv-install.
Executing: /lib/systemd/systemd-sysv-install enable docker
Loaded image: mirantis/hypokube:base

real	0m3.180s
user	0m0.196s
sys	0m0.136s

Step 1/2 : FROM mirantis/hypokube:base
 ---> a9b35ecee630
Step 2/2 : COPY hyperkube /hyperkube
 ---> 7252dd5583d3
Removing intermediate container fc3f9d3a23d0
Successfully built 7252dd5583d3
Created symlink /etc/systemd/system/multi-user.target.wants/kubelet.service → /lib/systemd/system/kubelet.service.
Flag --skip-preflight-checks has been deprecated, it is now equivalent to --ignore-preflight-errors=all
I0724 11:33:39.083574     481 feature_gate.go:230] feature gates: &{map[]}
[init] using Kubernetes version: v1.11.0
[preflight] running pre-flight checks
	[WARNING FileContent--proc-sys-net-bridge-bridge-nf-call-iptables]: /proc/sys/net/bridge/bridge-nf-call-iptables does not exist
	[WARNING FileExisting-crictl]: crictl not found in system path
I0724 11:33:39.112771     481 kernel_validator.go:81] Validating kernel version
I0724 11:33:39.112854     481 kernel_validator.go:96] Validating kernel config
[preflight/images] Pulling images required for setting up a Kubernetes cluster
[preflight/images] This might take a minute or two, depending on the speed of your internet connection
[preflight/images] You can also perform this action in beforehand using 'kubeadm config images pull'
[kubelet] Writing kubelet environment file with flags to file "/var/lib/kubelet/kubeadm-flags.env"
[kubelet] Writing kubelet configuration to file "/var/lib/kubelet/config.yaml"
[preflight] Activating the kubelet service
[certificates] Generated ca certificate and key.
[certificates] Generated apiserver certificate and key.
[certificates] apiserver serving cert is signed for DNS names [kube-master kubernetes kubernetes.default kubernetes.default.svc kubernetes.default.svc.cluster.local] and IPs [10.96.0.1 10.192.0.2]
[certificates] Generated apiserver-kubelet-client certificate and key.
[certificates] Generated sa key and public key.
[certificates] Generated front-proxy-ca certificate and key.
[certificates] Generated front-proxy-client certificate and key.
[certificates] Generated etcd/ca certificate and key.
[certificates] Generated etcd/server certificate and key.
[certificates] etcd/server serving cert is signed for DNS names [kube-master localhost] and IPs [127.0.0.1 ::1]
[certificates] Generated etcd/peer certificate and key.
[certificates] etcd/peer serving cert is signed for DNS names [kube-master localhost] and IPs [10.192.0.2 127.0.0.1 ::1]
[certificates] Generated etcd/healthcheck-client certificate and key.
[certificates] Generated apiserver-etcd-client certificate and key.
[certificates] valid certificates and keys now exist in "/etc/kubernetes/pki"
[kubeconfig] Wrote KubeConfig file to disk: "/etc/kubernetes/admin.conf"
[kubeconfig] Wrote KubeConfig file to disk: "/etc/kubernetes/kubelet.conf"
[kubeconfig] Wrote KubeConfig file to disk: "/etc/kubernetes/controller-manager.conf"
[kubeconfig] Wrote KubeConfig file to disk: "/etc/kubernetes/scheduler.conf"
[controlplane] wrote Static Pod manifest for component kube-apiserver to "/etc/kubernetes/manifests/kube-apiserver.yaml"
[controlplane] wrote Static Pod manifest for component kube-controller-manager to "/etc/kubernetes/manifests/kube-controller-manager.yaml"
[controlplane] wrote Static Pod manifest for component kube-scheduler to "/etc/kubernetes/manifests/kube-scheduler.yaml"
[etcd] Wrote Static Pod manifest for a local etcd instance to "/etc/kubernetes/manifests/etcd.yaml"
[init] waiting for the kubelet to boot up the control plane as Static Pods from directory "/etc/kubernetes/manifests" 
[init] this might take a minute or longer if the control plane images have to be pulled
[apiclient] All control plane components are healthy after 39.501516 seconds
[uploadconfig] storing the configuration used in ConfigMap "kubeadm-config" in the "kube-system" Namespace
[kubelet] Creating a ConfigMap "kubelet-config-1.11" in namespace kube-system with the configuration for the kubelets in the cluster
[markmaster] Marking the node kube-master as master by adding the label "node-role.kubernetes.io/master=''"
[markmaster] Marking the node kube-master as master by adding the taints [node-role.kubernetes.io/master:NoSchedule]
[patchnode] Uploading the CRI Socket information "/var/run/dockershim.sock" to the Node API object "kube-master" as an annotation
[bootstraptoken] using token: 7j7z69.8ulp428plm6sribh
[bootstraptoken] configured RBAC rules to allow Node Bootstrap tokens to post CSRs in order for nodes to get long term certificate credentials
[bootstraptoken] configured RBAC rules to allow the csrapprover controller automatically approve CSRs from a Node Bootstrap Token
[bootstraptoken] configured RBAC rules to allow certificate rotation for all node client certificates in the cluster
[bootstraptoken] creating the "cluster-info" ConfigMap in the "kube-public" namespace
[addons] Applied essential addon: kube-dns
[addons] Applied essential addon: kube-proxy

Your Kubernetes master has initialized successfully!

To start using your cluster, you need to run the following as a regular user:

  mkdir -p $HOME/.kube
  sudo cp -i /etc/kubernetes/admin.conf $HOME/.kube/config
  sudo chown $(id -u):$(id -g) $HOME/.kube/config

You should now deploy a pod network to the cluster.
Run "kubectl apply -f [podnetwork].yaml" with one of the options listed at:
  https://kubernetes.io/docs/concepts/cluster-administration/addons/

You can now join any number of machines by running the following on each node
as root:

  kubeadm join 10.192.0.2:6443 --token 7j7z69.8ulp428plm6sribh --discovery-token-ca-cert-hash sha256:b0fdb998586f1346624cd4cd6abbef2b77ed14e0610de99e37319554d4c606c7


real	0m56.444s
user	0m5.548s
sys	0m0.144s
999bc97a8c79
e1c9ecf0173b
642f4129fd91
ca567c3b9f9d
e70097978f0f
81b7120bfbe8
10043dfd2265
b8281a2d7c5c
Warning: kubectl apply should be used on resource created by either kubectl create --save-config or kubectl apply
configmap/kube-proxy configured
Warning: kubectl apply should be used on resource created by either kubectl create --save-config or kubectl apply
daemonset.extensions/kube-proxy configured
No resources found
* Setting cluster config 
Cluster "dind" set.
Context "dind" created.
* Starting node container: 1
* Starting DIND container: kube-node-1
* Node container started: 1
* Starting node container: 2
* Starting DIND container: kube-node-2
* Node container started: 2
* Joining node: 1
* Joining node: 2
* Running kubeadm: join --skip-preflight-checks 10.192.0.2:6443 --token 7j7z69.8ulp428plm6sribh --discovery-token-ca-cert-hash sha256:b0fdb998586f1346624cd4cd6abbef2b77ed14e0610de99e37319554d4c606c7
Initializing machine ID from random generator.
* Running kubeadm: join --skip-preflight-checks 10.192.0.2:6443 --token 7j7z69.8ulp428plm6sribh --discovery-token-ca-cert-hash sha256:b0fdb998586f1346624cd4cd6abbef2b77ed14e0610de99e37319554d4c606c7
Initializing machine ID from random generator.
Synchronizing state of docker.service with SysV service script with /lib/systemd/systemd-sysv-install.
Executing: /lib/systemd/systemd-sysv-install enable docker
Synchronizing state of docker.service with SysV service script with /lib/systemd/systemd-sysv-install.
Executing: /lib/systemd/systemd-sysv-install enable docker
Loaded image: mirantis/hypokube:base

real	0m5.143s
user	0m0.208s
sys	0m0.132s
Loaded image: mirantis/hypokube:base

real	0m5.924s
user	0m0.200s
sys	0m0.144s

Step 1/2 : FROM mirantis/hypokube:base
 ---> a9b35ecee630
Step 2/2 : COPY hyperkube /hyperkube

Step 1/2 : FROM mirantis/hypokube:base
 ---> a9b35ecee630
Step 2/2 : COPY hyperkube /hyperkube
 ---> c163693f4005
 ---> a421de292de6
Removing intermediate container 66e43f30b1d1
Successfully built c163693f4005
Removing intermediate container ebead67d8db5
Successfully built a421de292de6
Created symlink /etc/systemd/system/multi-user.target.wants/kubelet.service → /lib/systemd/system/kubelet.service.
Created symlink /etc/systemd/system/multi-user.target.wants/kubelet.service → /lib/systemd/system/kubelet.service.
Flag --skip-preflight-checks has been deprecated, it is now equivalent to --ignore-preflight-errors=all
[preflight] running pre-flight checks
	[WARNING RequiredIPVSKernelModulesAvailable]: the IPVS proxier will not be used, because the following required kernel modules are not loaded: [ip_vs ip_vs_rr ip_vs_wrr ip_vs_sh] or no builtin kernel ipvs support: map[ip_vs_wrr:{} ip_vs_sh:{} nf_conntrack_ipv4:{} ip_vs:{} ip_vs_rr:{}]
you can solve this problem with following methods:
 1. Run 'modprobe -- ' to load missing kernel modules;
2. Provide the missing builtin kernel ipvs support

	[WARNING FileContent--proc-sys-net-bridge-bridge-nf-call-iptables]: /proc/sys/net/bridge/bridge-nf-call-iptables does not exist
	[WARNING FileExisting-crictl]: crictl not found in system path
I0724 11:35:45.353641     463 kernel_validator.go:81] Validating kernel version
I0724 11:35:45.353750     463 kernel_validator.go:96] Validating kernel config
Flag --skip-preflight-checks has been deprecated, it is now equivalent to --ignore-preflight-errors=all
[preflight] running pre-flight checks
	[WARNING RequiredIPVSKernelModulesAvailable]: the IPVS proxier will not be used, because the following required kernel modules are not loaded: [ip_vs_sh ip_vs ip_vs_rr ip_vs_wrr] or no builtin kernel ipvs support: map[ip_vs:{} ip_vs_rr:{} ip_vs_wrr:{} ip_vs_sh:{} nf_conntrack_ipv4:{}]
you can solve this problem with following methods:
 1. Run 'modprobe -- ' to load missing kernel modules;
2. Provide the missing builtin kernel ipvs support

	[WARNING FileContent--proc-sys-net-bridge-bridge-nf-call-iptables]: /proc/sys/net/bridge/bridge-nf-call-iptables does not exist
	[WARNING FileExisting-crictl]: crictl not found in system path
I0724 11:35:45.473095     482 kernel_validator.go:81] Validating kernel version
I0724 11:35:45.473459     482 kernel_validator.go:96] Validating kernel config
[discovery] Trying to connect to API Server "10.192.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://10.192.0.2:6443"
[discovery] Trying to connect to API Server "10.192.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://10.192.0.2:6443"
[discovery] Failed to connect to API Server "10.192.0.2:6443": token id "7j7z69" is invalid for this cluster or it has expired. Use "kubeadm token create" on the master node to creating a new valid token
[discovery] Failed to connect to API Server "10.192.0.2:6443": token id "7j7z69" is invalid for this cluster or it has expired. Use "kubeadm token create" on the master node to creating a new valid token
[discovery] Trying to connect to API Server "10.192.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://10.192.0.2:6443"
[discovery] Trying to connect to API Server "10.192.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://10.192.0.2:6443"
[discovery] Requesting info from "https://10.192.0.2:6443" again to validate TLS against the pinned public key
[discovery] Requesting info from "https://10.192.0.2:6443" again to validate TLS against the pinned public key
[discovery] Cluster info signature and contents are valid and TLS certificate validates against pinned roots, will use API Server "10.192.0.2:6443"
[discovery] Successfully established connection with API Server "10.192.0.2:6443"
[discovery] Cluster info signature and contents are valid and TLS certificate validates against pinned roots, will use API Server "10.192.0.2:6443"
[discovery] Successfully established connection with API Server "10.192.0.2:6443"
[kubelet] Downloading configuration for the kubelet from the "kubelet-config-1.11" ConfigMap in the kube-system namespace
[kubelet] Downloading configuration for the kubelet from the "kubelet-config-1.11" ConfigMap in the kube-system namespace
[kubelet] Writing kubelet configuration to file "/var/lib/kubelet/config.yaml"
[kubelet] Writing kubelet configuration to file "/var/lib/kubelet/config.yaml"
[kubelet] Writing kubelet environment file with flags to file "/var/lib/kubelet/kubeadm-flags.env"
[preflight] Activating the kubelet service
[kubelet] Writing kubelet environment file with flags to file "/var/lib/kubelet/kubeadm-flags.env"
[preflight] Activating the kubelet service
[tlsbootstrap] Waiting for the kubelet to perform the TLS Bootstrap...
[tlsbootstrap] Waiting for the kubelet to perform the TLS Bootstrap...
[patchnode] Uploading the CRI Socket information "/var/run/dockershim.sock" to the Node API object "kube-node-2" as an annotation
[patchnode] Uploading the CRI Socket information "/var/run/dockershim.sock" to the Node API object "kube-node-1" as an annotation

This node has joined the cluster:
* Certificate signing request was sent to master and a response
  was received.
* The Kubelet was informed of the new secure connection details.

Run 'kubectl get nodes' on the master to see this node join the cluster.

real	0m13.949s
user	0m0.616s
sys	0m0.064s
* Node joined: 1

This node has joined the cluster:
* Certificate signing request was sent to master and a response
  was received.
* The Kubelet was informed of the new secure connection details.

Run 'kubectl get nodes' on the master to see this node join the cluster.

real	0m14.094s
user	0m0.604s
sys	0m0.068s
* Node joined: 2
Creating static routes for bridge plugin
* Deploying k8s dashboard 
deployment.extensions/kubernetes-dashboard created
service/kubernetes-dashboard created
clusterrolebinding.rbac.authorization.k8s.io/add-on-cluster-admin created
* Patching kube-dns deployment to make it start faster 
Warning: kubectl apply should be used on resource created by either kubectl create --save-config or kubectl apply
deployment.extensions/kube-dns configured
* Taking snapshot of the cluster 
deployment.extensions/kube-dns scaled
deployment.extensions/kubernetes-dashboard scaled
pod "kube-proxy-g68lz" deleted
pod "kube-proxy-r6dbq" deleted
pod "kube-proxy-zfcg7" deleted
NAME                        READY     STATUS              RESTARTS   AGE
kube-dns-86c47599bd-6lhbd   2/3       Terminating         0          32s
kube-proxy-6ng2g            1/1       Running             0          6s
kube-proxy-dlqsb            0/1       ContainerCreating   0          1s
kube-proxy-wl827            1/1       Running             0          2s
tar: var/lib/kubelet/device-plugins/kubelet.sock: socket ignored
tar: var/lib/kubelet/device-plugins/kubelet.sock: socket ignored
Warning: Stopping docker.service, but it can still be activated by:
  docker.socket
tar: var/lib/kubelet/device-plugins/kubelet.sock: socket ignored
* Waiting for kube-proxy and the nodes 
....[done]
* Bringing up kube-dns and kubernetes-dashboard 
deployment.extensions/kube-dns scaled
deployment.extensions/kubernetes-dashboard scaled
..........................[done]
NAME          STATUS    ROLES     AGE       VERSION
kube-master   Ready     master    3m        v1.11.0
kube-node-1   Ready     <none>    1m        v1.11.0
kube-node-2   Ready     <none>    1m        v1.11.0
* Access dashboard at: http://localhost:8080/api/v1/namespaces/kube-system/services/kubernetes-dashboard:/proxy
export PATH="/home/travis/.kubeadm-dind-cluster:/home/travis/gopath/bin:/home/travis/.gimme/versions/go1.10.linux.amd64/bin:/home/travis/bin:/home/travis/bin:/home/travis/.local/bin:/opt/pyenv/shims:/home/travis/.phpenv/shims:/home/travis/perl5/perlbrew/bin:/home/travis/.nvm/versions/node/v8.9.1/bin:/home/travis/.kiex/elixirs/elixir-1.4.5/bin:/home/travis/.kiex/bin:/home/travis/.rvm/gems/ruby-2.4.1/bin:/home/travis/.rvm/gems/ruby-2.4.1@global/bin:/home/travis/.rvm/rubies/ruby-2.4.1/bin:/home/travis/gopath/bin:/home/travis/.gimme/versions/go1.7.4.linux.amd64/bin:/usr/local/phantomjs/bin:/usr/local/phantomjs:/usr/local/neo4j-3.2.7/bin:/usr/local/maven-3.5.2/bin:/usr/local/cmake-3.9.2/bin:/usr/local/clang-5.0.0/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/home/travis/.rvm/bin:/home/travis/.phpenv/bin:/opt/pyenv/bin:/home/travis/.yarn/bin"
nohup _output/bin/kar-controllers --master localhost:8080 --logtostderr --v 3 > controller.log 2>&1 &
nohup _output/bin/kar-scheduler --master localhost:8080 --logtostderr --v 3 > scheduler.log 2>&1 &
# kubectl config set current-context dind
cat /home/travis/.kube/config
apiVersion: v1
clusters:
- cluster:
    insecure-skip-tls-verify: true
    server: http://localhost:8080
  name: dind
contexts:
- context:
    cluster: dind
    user: ""
  name: dind
current-context: ""
kind: Config
preferences: {}
users: []
cd test && go test -v
=== RUN   TestE2E
Running Suite: kube-arbitrator Test Suite
=========================================
Random Seed: 1532432269
Will run 5 of 5 specs

• Failure [0.002 seconds]
E2E Test
/home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:24
  Schedule QueueJob with SchedulerName [It]
  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:25

  Expected error:
      <clientcmd.errConfigurationInvalid | len:1, cap:1>: [
          {
              s: "no configuration has been provided",
          },
      ]
      invalid configuration: no configuration has been provided
  not to have occurred

  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/util.go:76
------------------------------
• Failure [0.002 seconds]
E2E Test
/home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:24
  Schedule QueueJob [It]
  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:34

  Expected error:
      <clientcmd.errConfigurationInvalid | len:1, cap:1>: [
          {
              s: "no configuration has been provided",
          },
      ]
      invalid configuration: no configuration has been provided
  not to have occurred

  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/util.go:76
------------------------------
• Failure [0.002 seconds]
E2E Test
/home/travislid | len:1, cap:1>: [
          {
              s: "no configuration has been provided",
          },
      ]
      invalid configuration: no configuration has been provided
  not to have occurred

  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/util.go:76
------------------------------


Summarizing 5 Failures:

[Fail] E2E Test [It] Schedule QueueJob with SchedulerName 
…